+++ query_result_document.py
@@ -1,5 +1,8 @@
+from numpy import uint16
+
+
 class QueryResultDocument:
   def __init__(self, result_list, rank, document):
     self.result_list = result_list
-    self.rank = rank
+    self.rank = uint16(rank)
     self.document = document
